backend/requirements.txt
setuptools>=70.0.0
wheel>=0.43.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
audioop-lts==0.2.2
audioread==3.0.1
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
elevenlabs==2.16.0
fastapi==0.118.3
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.0
MarkupSafe==3.0.3
matplotlib==3.10.6
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
narwhals==2.7.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
oauthlib==3.3.1
openai==1.109.1
openai-whisper==20250625
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pooch==1.8.2
proglog==0.1.12
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.48.0
streamlit==1.50.0
streamlit-autorefresh==1.0.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
